refactor: separate Dxtrade and MT5 CreateAccountButton
